zlh 代码优化

+++ src/main/java/com/hc/controller/task/TaskEmailController.java
@@ -7,12 +7,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.hc.common.exception.CustomException;
 import com.hc.common.result.ResultBase;
 import com.hc.para.page_base.BasePara;
 import com.hc.pojo.task.TaskInfo;
 import com.hc.service.TaskService;
+import com.hc.utils.result.ResultUtil;
 
 @Controller
 @RequestMapping("/email/task/")
@@ -38,8 +40,8 @@ public class TaskEmailController {
 	 * @param info
 	 */
 	@RequestMapping(value = "save")
-	public ResultBase save(@RequestBody(required = false) TaskInfo info) throws Exception, CustomException {
-		return taskService.addJob(info, 0);
+	public ResultBase save(TaskInfo info,MultipartFile files) throws Exception, CustomException {
+		return taskService.addJob(info, 0 ,files);
 	}
 
 	/**
@@ -60,6 +62,9 @@ public class TaskEmailController {
 	 */
 	@RequestMapping(value = "delete")
 	public ResultBase delete(@RequestBody(required = false) TaskInfo info) throws Exception, CustomException {
+		if(info==null) {
+			return ResultUtil.getResultBase(false);
+		}
 		return taskService.delete(info.getJobName(), info.getJobGroup());
 	}
